WIP: Select multiple properties
The start of work to allow multiple properties. The issue has more to do
with how objects and arrays are started and ended, and the scope writing
would have to be revamped. As this work isn't necessary to optionally
emit packages, I'm stashing it away on the server for now and may
revisit later.